refactor frontend components
