+++ src/frontend/src/components/home/index.ts
@@ -1,0 +1,4 @@
+export { default as HeroSection } from './HeroSection';
+export { default as HowItWorks } from './HowItWorks';
+export { default as Benefits } from './Benefits';
+export { default as FinalCTA } from './FinalCTA';
